requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
idna==3.4
kiwisolver==1.4.5
load-dotenv==0.1.0
matplotlib==3.8.1
mplfinance==0.12.10b0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
